# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PMS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Drug": {
            "post": {
                "tags": [
                    "Drug"
                ],
                "summary": "Create new Drug",
                "operationId": "createDrug",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Drug name",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "generic": {
                                        "description": "Drug generic name",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "description": {
                                        "description": "Drug description",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "price": {
                                        "description": "Drug price",
                                        "type": "integer",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrugSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drug/{id}": {
            "get": {
                "tags": [
                    "Drug"
                ],
                "summary": "Find Drug by id",
                "operationId": "getDrug",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Drug ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrugSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Drug not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Drug"
                ],
                "summary": "Edit Drug by id",
                "operationId": "editDrug",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Drug ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Drug name",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "generic": {
                                        "description": "Drug generic name",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "description": {
                                        "description": "Drug description",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "price": {
                                        "description": "Drug price",
                                        "type": "integer",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrugSchema"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Drug"
                ],
                "summary": "Delete Drug by id",
                "operationId": "deleteDrug",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Drug ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "402": {
                        "description": "Drug not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DrugSchema": {
                "properties": {
                    "id": {
                        "description": "Drug identifier",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "Drug name",
                        "type": "string",
                        "nullable": "false"
                    },
                    "generic": {
                        "description": "Drug generic name",
                        "type": "string",
                        "nullable": "false"
                    },
                    "description": {
                        "description": "Drug description",
                        "type": "string",
                        "nullable": "false"
                    },
                    "price": {
                        "description": "Drug price",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "pharmacy": {
                        "$ref": "#/components/schemas/PharmacySchema"
                    }
                },
                "type": "object"
            },
...
            "PharmacySchema": {
                "properties": {
                    "id": {
                        "description": "Pharmacy identifier",
                        "type": "integer",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "Pharmacy name",
                        "type": "string",
                        "nullable": "false"
                    },
                    "licence": {
                        "description": "Pharmacy licence id",
                        "type": "string",
                        "nullable": "false"
                    },
                    "safe": {
                        "description": "Safe amount",
                        "type": "number",
                        "format": "double",
                        "nullable": "false"
                    },
                    "chest": {
                        "description": "Chest amount",
                        "type": "number",
                        "format": "double",
                        "nullable": "false"
                    }
                },
                "type": "object"
            }
        },
...
    }
}